+++ examples/docs/src/pages/_app.js
@@ -1,6 +1,0 @@
-import 'nextra-theme-docs/style.css'
-
-export default function Nextra({ Component, pageProps }) {
-  const getLayout = Component.getLayout || (page => page)
-  return getLayout(<Component {...pageProps} />)
-}