+++ packages/nextra-theme-docs/src/components/bleed.tsx
@@ -11,11 +11,11 @@ export function Bleed({
   return (
     <div
       className={cn(
-        'bleed relative -mx-6 mt-6 md:-mx-8 2xl:-mx-24',
+        'bleed nx-relative -nx-mx-6 nx-mt-6 md:-nx-mx-8 2xl:-nx-mx-24',
         full && [
-          'md:mx:[calc(-50vw+50%+8rem)',
-          'ltr:xl:ml-[calc(50%-50vw+16rem)] ltr:xl:mr-[calc(50%-50vw)]',
-          'rtl:xl:ml-[calc(50%-50vw)] rtl:xl:mr-[calc(50%-50vw+16rem)]'
+          // 'md:mx:[calc(-50vw+50%+8rem)',
+          'ltr:xl:nx-ml-[calc(50%-50vw+16rem)] ltr:xl:nx-mr-[calc(50%-50vw)]',
+          'rtl:xl:nx-ml-[calc(50%-50vw)] rtl:xl:nx-mr-[calc(50%-50vw+16rem)]'
         ]
       )}
     >