+++ packages/nextra-theme-docs/src/components/theme-switch.tsx
@@ -19,25 +19,25 @@ export function ThemeSwitch({ lite }: ThemeSwitchProps): ReactElement {
   const mounted = useMounted()
   const IconToUse = mounted && resolvedTheme === 'dark' ? MoonIcon : SunIcon
   return (
-  <div className="relative">
-    <Select
-      title="Change theme"
-      options={OPTIONS}
-      onChange={option => {
-        setTheme(option.key)
-      }}
-      selected={{
-        key: theme,
-        name: (
-          <div className="flex items-center gap-2 capitalize">
-            <IconToUse />
-            <span className={lite ? 'md:hidden' : ''}>
-              {mounted ? theme : 'light'}
-            </span>
-          </div>
-        )
-      }}
-    />
-  </div>
+    <div className="nx-relative">
+      <Select
+        title="Change theme"
+        options={OPTIONS}
+        onChange={option => {
+          setTheme(option.key)
+        }}
+        selected={{
+          key: theme,
+          name: (
+            <div className="nx-flex nx-items-center nx-gap-2 nx-capitalize">
+              <IconToUse />
+              <span className={lite ? 'md:nx-hidden' : ''}>
+                {mounted ? theme : 'light'}
+              </span>
+            </div>
+          )
+        }}
+      />
+    </div>
   )
 }