+++ packages/nextra-theme-docs/src/constants.tsx
@@ -15,19 +15,16 @@ export const IS_BROWSER = typeof window !== 'undefined'
 export const DEFAULT_THEME: DocsThemeConfig = {
   banner: {
     dismissible: true,
-    key: 'nextra-banner',
-    text: ''
+    key: 'nextra-banner'
   },
   chat: {
     icon: (
       <>
         <DiscordIcon />
-        <span className="sr-only">Discord</span>
+        <span className="nx-sr-only">Discord</span>
       </>
-    ),
-    link: ''
+    )
   },
-  components: {},
   darkMode: true,
   direction: 'ltr',
   docsRepositoryBase: 'https://github.com/shuding/nextra',
@@ -45,15 +42,12 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     },
     text: 'Edit this page'
   },
-  faviconGlyph: '',
-  feedback: {
-    content: null,
-    labels: ''
-  },
+  feedback: {},
   footer: {
     component: Footer,
     text: `MIT ${new Date().getFullYear()} © Nextra.`
   },
+  getNextSeoProps: () => ({ titleTemplate: '%s – Nextra' }),
   gitTimestamp({ timestamp }) {
     const { locale = DEFAULT_LOCALE } = useRouter()
     return (
@@ -89,9 +83,6 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     </>
   ),
   logoLink: true,
-  main: {
-    extraContent: null
-  },
   navbar: {
     component: Navbar,
     extraContent: null
@@ -116,11 +107,9 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     icon: (
       <>
         <GitHubIcon />
-        <span className="sr-only">GitHub</span>
+        <span className="nx-sr-only">GitHub</span>
       </>
-    ),
-    // by default should be empty so clicking on project link will go to the github link
-    link: ''
+    )
   },
   search: {
     component({ className, directories }) {
@@ -138,8 +127,10 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     ),
     placeholder() {
       const { locale } = useRouter()
-      if (locale === 'zh-CN') return '搜索文档...'
-      return 'Search documentation...'
+      if (locale === 'zh-CN') return '搜索文档…'
+      if (locale === 'ru-RU') return 'Поиск документации…'
+      if (locale === 'fr-FR') return 'Rechercher de la documentation…'
+      return 'Search documentation…'
     }
   },
   serverSideError: {
@@ -150,10 +141,8 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     defaultMenuCollapsed: false,
     titleComponent: ({ title }) => <>{title}</>
   },
-  titleSuffix: ' – Nextra',
   toc: {
     component: TOC,
-    extraContent: null,
     float: true,
     title: 'On This Page'
   }
@@ -174,7 +163,7 @@ export const DEEP_OBJECT_KEYS = Object.entries(DEFAULT_THEME)
 
 export const LEGACY_CONFIG_OPTIONS: Record<string, string> = {
   bannerKey: 'banner.key',
-  bodyExtraContent: 'main.extraContent',
+  bodyExtraContent: 'main',
   customSearch: 'search.component',
   defaultMenuCollapsed: 'sidebar.defaultMenuCollapsed',
   feedbackLabels: 'feedback.labels',