+++ packages/nextra-theme-docs/src/mdx-components.tsx
@@ -108,19 +108,19 @@ const createHeaderLink = (
     return (
       <Tag
         className={cn(
-          'font-semibold tracking-tight',
+          'nx-font-semibold nx-tracking-tight',
           {
-            h2: 'mt-10 text-3xl border-b pb-1 dark:border-primary-100/10 contrast-more:border-neutral-400 contrast-more:dark:border-neutral-400',
-            h3: 'mt-8 text-2xl',
-            h4: 'mt-8 text-xl',
-            h5: 'mt-8 text-lg',
-            h6: 'mt-8 text-base'
+            h2: 'nx-mt-10 nx-text-3xl nx-border-b nx-pb-1 dark:nx-border-primary-100/10 contrast-more:nx-border-neutral-400 contrast-more:dark:nx-border-neutral-400',
+            h3: 'nx-mt-8 nx-text-2xl',
+            h4: 'nx-mt-8 nx-text-xl',
+            h5: 'nx-mt-8 nx-text-lg',
+            h6: 'nx-mt-8 nx-text-base'
           }[Tag]
         )}
         {...props}
       >
         {children}
-        <span className="absolute -mt-20" id={id} ref={obRef} />
+        <span className="nx-absolute -nx-mt-20" id={id} ref={obRef} />
         <a href={`#${id}`} className="subheading-anchor" />
       </Tag>
     )
@@ -180,7 +180,7 @@ const Details = ({
 
   return (
     <details
-      className="my-4 rounded border border-gray-200 bg-white p-2 shadow-sm dark:border-neutral-800 dark:bg-neutral-900 first:mt-0 last:mb-0"
+      className="nx-my-4 nx-rounded nx-border nx-border-gray-200 nx-bg-white nx-p-2 nx-shadow-sm dark:nx-border-neutral-800 dark:nx-bg-neutral-900 first:nx-mt-0 last:nx-mb-0"
       {...props}
       open={delayedOpenState}
       {...(openState && { 'data-expanded': true })}
@@ -196,9 +196,9 @@ const Summary = (props: ComponentProps<'summary'>): ReactElement => {
   return (
     <summary
       className={cn(
-        'list-none cursor-pointer p-1 transition-colors hover:bg-gray-100 dark:hover:bg-neutral-800',
-        "before:mr-1 before:content-[''] before:inline-block before:transition-transform dark:before:invert",
-        '[[data-expanded]>&]:before:rotate-90 rtl:before:rotate-180'
+        'nx-list-none nx-cursor-pointer nx-p-1 nx-transition-colors hover:nx-bg-gray-100 dark:hover:nx-bg-neutral-800',
+        "before:nx-mr-1 before:content-[''] before:nx-inline-block before:nx-transition-transform dark:before:nx-invert",
+        '[[data-expanded]>&]:before:nx-rotate-90 rtl:before:nx-rotate-180'
       )}
       {...props}
       onClick={e => {
@@ -227,7 +227,10 @@ export const getComponents = ({
   const context = { index: 0 }
   return {
     h1: (props: ComponentProps<'h1'>) => (
-      <h1 className="mt-2 text-4xl font-bold tracking-tight" {...props} />
+      <h1
+        className="nx-mt-2 nx-text-4xl nx-font-bold nx-tracking-tight"
+        {...props}
+      />
     ),
     h2: createHeaderLink('h2', context),
     h3: createHeaderLink('h3', context),
@@ -235,35 +238,44 @@ export const getComponents = ({
     h5: createHeaderLink('h5', context),
     h6: createHeaderLink('h6', context),
     ul: (props: ComponentProps<'ul'>) => (
-      <ul className="ltr:ml-6 rtl:mr-6 mt-6 list-disc first:mt-0" {...props} />
+      <ul
+        className="ltr:nx-ml-6 rtl:nx-mr-6 nx-mt-6 nx-list-disc first:nx-mt-0"
+        {...props}
+      />
     ),
     ol: (props: ComponentProps<'ol'>) => (
-      <ol className="ltr:ml-6 rtl:mr-6 mt-6 list-decimal" {...props} />
+      <ol
+        className="ltr:nx-ml-6 rtl:nx-mr-6 nx-mt-6 nx-list-decimal"
+        {...props}
+      />
     ),
-    li: (props: ComponentProps<'li'>) => <li className="my-2" {...props} />,
+    li: (props: ComponentProps<'li'>) => <li className="nx-my-2" {...props} />,
     blockquote: (props: ComponentProps<'blockquote'>) => (
       <blockquote
         className={cn(
-          'mt-6 first:mt-0 border-gray-300 italic text-gray-700 dark:border-gray-700 dark:text-gray-400',
-          'rtl:border-r-2 rtl:pr-6 ltr:border-l-2 ltr:pl-6'
+          'nx-mt-6 first:nx-mt-0 nx-border-gray-300 nx-italic nx-text-gray-700 dark:nx-border-gray-700 dark:nx-text-gray-400',
+          'rtl:nx-border-r-2 rtl:nx-pr-6 ltr:nx-border-l-2 ltr:nx-pl-6'
         )}
         {...props}
       />
     ),
     hr: (props: ComponentProps<'hr'>) => (
-      <hr className="my-8 dark:border-gray-900" {...props} />
+      <hr className="nx-my-8 dark:nx-border-gray-900" {...props} />
     ),
     a: props => (
       <A
         {...props}
-        className="text-primary-500 underline decoration-from-font [text-underline-position:under]"
+        className="nx-text-primary-500 nx-underline nx-decoration-from-font [text-underline-position:under]"
       />
     ),
     table: (props: ComponentProps<'table'>) => (
-      <Table className="nextra-scrollbar mt-6 first:mt-0 p-0" {...props} />
+      <Table
+        className="nextra-scrollbar nx-mt-6 first:nx-mt-0 nx-p-0"
+        {...props}
+      />
     ),
     p: (props: ComponentProps<'p'>) => (
-      <p className="mt-6 first:mt-0 leading-7" {...props} />
+      <p className="nx-mt-6 first:nx-mt-0 nx-leading-7" {...props} />
     ),
     tr: Tr,
     th: Th,